+++ src/textual/drivers/linux_inline_driver.py
@@ -266,6 +266,10 @@ class LinuxInlineDriver(Driver):
             # TODO: log this
             pass
 
+    def flush(self):
+        """Flush any buffered data."""
+        self._file.flush()
+
     def stop_application_mode(self) -> None:
         """Stop application mode, restore state."""
         self._disable_bracketed_paste()